Move CultureGroup names into its own file
