+++ Vic2ToHoI4/Vic2ToHoI4.vcxproj.filters
@@ -818,6 +818,9 @@
     <ClCompile Include="Source\HOI4World\Names\CultureNames.cpp">
       <Filter>HoI4World\Names</Filter>
     </ClCompile>
+    <ClCompile Include="Source\HOI4World\Names\CultureGroupNames.cpp">
+      <Filter>HoI4World\Names</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\common_items\Date.h">
@@ -1539,6 +1542,9 @@
       <Filter>HoI4World\Names</Filter>
     </ClInclude>
     <ClInclude Include="Source\HOI4World\Names\CultureNames.h">
+      <Filter>HoI4World\Names</Filter>
+    </ClInclude>
+    <ClInclude Include="Source\HOI4World\Names\CultureGroupNames.h">
       <Filter>HoI4World\Names</Filter>
     </ClInclude>
   </ItemGroup>
